Refactor into a package
Note that Python has an unfortunate naming collision here. This commit
refactors the existing pshtt code into a "package" - that is, a
multi-file module with an `__init__.py`. This is distinct from a
"package on PyPi" (which is a "distribution package"). However, this
refactoring does make it easier to write a `setup.py` which can be used
to create a package for PyPi.

This refactor changes the calling convention for the pshtt CLI tool from
`./pshtt_cli` to `python -m pshtt.cli`. The Docker entrypoint script and
the README are updated accordingly.