Fix incorrect null results for "Downgrades HTTPS"
This commit fixes 2 bugs with the logic for computing "Downgrades HTTPS":

1. If a site **does not** have HTTPS (`"Valid HTTPS": false"`), then
"Downgrades HTTPS" should be `null` because it doesn't make sense to ask if a
site downgrades HTTPS if it doesn't have HTTPS in the first place. Previously,
the code would return `false` in this case.

2. If a site **does** have HTTPS, there were some cases where the previous code
would incorrectly return `null` instead of `false`. Specifically, this would
happen if a site's domain name was also the canonical endpoint, in which case
the site does not redirect and all values related to redirects are None. Due to
Python's short-circuit evaluation of boolean expressions, this None value was
returned instead of the intended False from `is_downgrades_https`.

Fixes #49.