Added more robust error handling
1) Add try/catches with debug output that shows where an unknown
exception occured.
2) Add an unknown_exception key to the endpoints to show where the
website ran into an unknown_exception for more debugging.
3) Add "Unknown Exception" at the top level to flag that something
in the website went wrong.
4) Add OpenSSL error handling to requests error handling for
more in depth error handling.

momimoun.com is an example of where hsts_check goes wrong
whoohoo.co.uk was the sole reason for creating the openssl exception
handling

I should note that I didn't include the error messages in the output but
rather in the debug output to keep consistency with pshtt.