Fixes bugs where the globals are None
This is the case when these functions are called directly, and where the
`inspect_domains()` function is not used.

Currently these lists are defined only if `inspect_domains()` has been
called, which may not be the case when testing (e.g. in #28 / #125), or
when using pshtt as a library. The current solution is to raise an
exception if the initialize function is not called explicitly, which is
now its own function to handle initializing all the third party data.

This begins work towards a more complete solution for #99, and allows
for the initialize function to be mocked / called separately when
working in tests.