Since we aren't interested in the actual content of the request, this code now uses the streaming variant of python-requests.  This variant delays the retrieval of the content until we access Response.content. This will:  * Save us time and bandwidth  * Stop pshtt from hanging on URLs that stream neverending data, like    webcams.  See #138:    https://github.com/dhs-ncats/pshtt/issues/138
Since we are not actually reading the content of the request we have
to be careful to ensure that the close() method is called on the
Request object returned by the ping method.  That is the ONLY way the
connection can be closed and released back into the pool.  One way to
ensure this happens is to use the "with" Python construct.