Version Pinning for "Operational" *master* Branch
Given the recent trouble with our dependencies drifting in and out of compatability, I'm proposing we version pin *master* to a conservative, known-to-be-working set. I took a naive approach here and just `==` the latest versions of various components that I knew worked for my personal usage. There may be a more pythonic way of doing this, but I'm not super familiar with the idioms.