Flexible Dependencies for Development
In addition to pinning dependencies in *master*, it might make sense to keep the *develop* branch nice and flexible, with the expectation that it might break.

I think it makes sense to retain the capability to build from "unpublished" source when necessary.

As a rule-of-thumb, maybe we also agree that the minimum version accepted by *develop* be the version we pin in *master*.