Tidy up logic
The original logic was confusing, because it relied on the fact that
and takes precedence over or in Python.  Originally I was going to add
a set of parentheses to give:
    return is_domain_supports_https(domain) and
        (
            is_strictly_forces_https(domain) and
            (
                is_defaults_to_https(domain) or
                is_redirect(domain)
            )
        ) or (
            (not is_strictly_forces_https(domain)) and
            is_defaults_to_https(domain)
        )
    )

After staring at this a bit, though, I realized that if
is_defaults_to_https(domain) is true then the value of
is_strictly_forces_https(domain) is irrelevant.  This led me to
realize that this expression can be simplified to:
    return is_domain_supports_https(domain) and (
        is_defaults_to_https(domain) or (
            is_strictly_forces_https(domain) and is_redirect(domain)
        )
    )