Configure flake8 to ignore W504
This is a warning for things like
a and
(b or c).

flake8 prefers
a and (
    b or c
)

The problem is that there are dozens of these warnings, and in many
cases they are in some complicated logic with LISPy levels of
parentheses.  "Fixing" these warnings is dangerous because it would be
easy to make flake8 happy but accidentally change the logic.