Fix test
Note that changing the logic from:
if domain.https.live == False and domain.httpswww.live == False:
to:
if not domain.https.live and not domain.httpswww.live:
changes the result when both domain.https.live and
domain.httpswww.live are both None.  This is because None != False but
not None is True.

I think that in the edge case where domain.https.live == None and
domain.httpswww.live == None it makes more sense to return "N/A" anyway.