Remove an extraneous call to is_http_redirect_domain(domain)
It's unclear to me why that function call is there, but it is not
needed.  When this code was last modified, the existing logic was:

    return is_domain_supports_https(domain) and (
        is_strictly_forces_https(domain) and
        (
            is_defaults_to_https(domain) or
            is_redirect(domain)
        ) or (
            (not is_strictly_forces_https(domain)) and
            is_defaults_to_https(domain)
        )
    )

With this bugfix, the logic is now:

    return is_domain_supports_https(domain) and (
        is_defaults_to_https(domain) or (
            is_strictly_forces_https(domain) and is_redirect_domain(domain)
        )
    )

A quick logic test using the eight possible values of
is_strictly_forces_https(domain), is_defaults_to_https(domain), and
is_redirect(domain) shows that the old and new expressions are now in
agreement.

This was noticed thanks to a comment by @climber-girl.