Temporarily pin sslyze to version 2.0.6
sslyze went through some breaking changes in version 2.1.1.  I am
working on adapting to those changed in #199, but those changes need
to be thoroughly tested before deployment.  In the interest of getting
the logging changes from #197 deployed for this weekend's scanning
run, I will temporarily pin the sslyze version to 2.0.6.  This pinning
will be removed once #199 is tested, approved, and merged.