Refactor the pshtt.utils.debug function
This function is something of a wrapper on the logging.debug()
function, but it forces calls to handle string formatting because of
how it is written. This refactor makes it a proper wrapper that only
handles printing the divider and passes through any positional
arguments to the underlying logging.debug() call.