Fix pre-commit configuration due to importlib-metadata 5
Update the flake8 hook to use the correct repository and update the
version to the latest tag available in that new repository. Add an
importlib-metadata pin to both bandit hooks as bandit is not currently
able to work with importlib-metadata 5. The status of this issue can be
tracked in PyCQA/bandit#956.