Install/upgrade setuptools and wheel when upgrading pip
When wheel gets installed alongside other packages, it may not get
used when those other packages are installed.  When that happens I see
warnings like this:
  DEPRECATION: ansible-core is being installed using the legacy
  'setup.py install' method, because it does not have a
  'pyproject.toml' and the 'wheel' package is not installed. pip 23.1
  will enforce this behaviour change. A possible replacement is to
  enable the '--use-pep517' option. Discussion can be found at
  https://github.com/pypa/pip/issues/8559

This change should get rid of these warnings.

Nota bene: This is the practice we follow in the Dockerfile in
cisagov/skeleton-docker, but for some reason we never started using it
in our workflows.